backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Security & Validation
bcrypt==4.1.2
marshmallow==3.20.2
python-dotenv==1.0.0

# File Handling
# Pillow==11.0.0  # Temporarily disabled due to Python 3.14 incompatibility
Werkzeug==3.0.1
cloudinary==1.40.0

# Production Server
gunicorn==21.2.0

# Testing
requests==2.31.0
